structure a folder and components
