+++ src/components/Vacancies/components/VacanсyHeader.tsx
@@ -1,0 +1,11 @@
+import { FC } from "react";
+
+const VacancyHeader: FC = () => {
+    return (
+        <div className="flex w-full items-center justify-center">
+            <h2 className="text-2xl">Vacancy filters and search</h2>
+        </div>
+    )
+}
+
+export default VacancyHeader;
